Add Film.ToSensorRGB, remove unused Camera1, tweak Cornell box
Film now converts a sampled spectrum + wavelengths to sensor RGB via
the CIE XYZ curves. Removed the dead Camera1 (superseded by the
Camera/ classes), brightened the Cornell box light, and fixed a
couple of implicit-cast warnings in the spectrum sampling code.

+++ Moarx.Graphics/Spectrum/SampledWavelengths.cs
@@ -22,7 +22,7 @@ public class SampledWavelengths {
         }
 
         for (var i = 0; i < ISpectrum.NSpectrumSamples; i++) {
-            swl.pdf[i] = 1 / (lambdaMax - lambdaMin);
+            swl.pdf[i] = (double)1 / (lambdaMax - lambdaMin);
         }
 
         return swl;
